requirements.txt
anthropic
fastapi
uvicorn[standard]
websockets
httpx
beautifulsoup4
tavily-python
chromadb
pydantic
pydantic-settings
python-dotenv
pytest
pytest-asyncio
gspread
google-auth